fix: restructure session implementation, add typed errors, rename session struct
