+++ token/services/ttx/receivetx.go
@@ -14,6 +14,7 @@ import (
 	"github.com/hyperledger-labs/fabric-smart-client/platform/view/view"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/services/utils"
 	jsession "github.com/hyperledger-labs/fabric-token-sdk/token/services/utils/json/session"
+	utilsession "github.com/hyperledger-labs/fabric-token-sdk/token/services/utils/session"
 )
 
 // ReceiveTransactionView is a view to read a transaction from the context's session.
@@ -49,7 +50,7 @@ func (f *ReceiveTransactionView) Call(context view.Context) (interface{}, error)
 	jsonSession := jsession.JSON(context)
 	msg, err := jsonSession.ReceiveRawWithTimeout(options.Timeout)
 	if err != nil {
-		if errors.Is(err, jsession.ErrTimeout) {
+		if errors.Is(err, utilsession.ErrTimeout) {
 			return nil, errors.Join(err, ErrTimeout)
 		}
 		logger.ErrorfContext(context.Context(), err.Error())
